Change all MEZAHE_RESHUMA* to long

+++ Edu.Mzt.Partners.Interfaces/Talmid/HatzharatMutav.cs
@@ -12,7 +12,7 @@ namespace Edu.Mzt.Partners.Interfaces.Talmid
         public int MISPAR_ZEHUT_TALMID { get; set; }
         public int SEMEL_MOSAD_TALMID { get; set; }
         public int SHNAT_LIMUDIM { get; set; }
-        public int MEZAHE_RESHUMA_SHIBUTZ { get; set; }
+        public long MEZAHE_RESHUMA_SHIBUTZ { get; set; }
         public int CODE_STATUS_HATZHARA_MUTAV { get; set; }
         public int MEZAHE_HATZHARA_MUTAV { get; set; }
         public DateTime TAARICH_IDKUN_MEZAHE_HATZHARA_MUTAV { get; set; }
